Move home components
